Move metrics_on_predictions out of experiments
It is used elsewhere, unlike the rest of the experiment utility code.
