+++ csrank/tests/test_choice_functions.py
@@ -8,7 +8,7 @@ from pymc3.variational.callbacks import CheckParametersConvergence
 
 from csrank.choicefunction import *
 from csrank.experiments.constants import *
-from csrank.experiments.util import metrics_on_predictions
+from csrank.util import metrics_on_predictions
 from csrank.metrics_np import f1_measure, subset_01_loss, instance_informedness, auc_score
 from csrank.tests.test_ranking import check_params_tunable
