+++ csrank/tests/test_discrete_choice.py
@@ -9,7 +9,7 @@ from pymc3.variational.callbacks import CheckParametersConvergence
 from csrank.dataset_reader.discretechoice.util import convert_to_label_encoding
 from csrank.discretechoice import *
 from csrank.experiments.constants import *
-from csrank.experiments.util import metrics_on_predictions
+from csrank.util import metrics_on_predictions
 from csrank.metrics_np import categorical_accuracy_np, topk_categorical_accuracy_np, subset_01_loss
 from csrank.tests.test_ranking import check_params_tunable
